requirements.txt
Django==4.1.2
djangorestframework==3.14.0
mysqlclient==2.2.0
loguru == 0.7.2
python-decouple==3.8
django-shell-plus==1.1.7
celery==5.1.1
redis==3.5.3



